refactor(components): [color-picker] alpha-slider (#10332)
* refactor(components): [color-picker] alpha-slider

* Extract logic code out of color-picker/alpha-slider

* fix: exposing methods

* chore: expose refs

---------
