+++ packages/components/color-picker/src/props/alpha-slider.ts
@@ -1,0 +1,17 @@
+import { buildProps, definePropType } from '@element-plus/utils'
+
+import type { ExtractPropTypes } from 'vue'
+import type Color from '../utils/color'
+
+export const alphaSliderProps = buildProps({
+  color: {
+    type: definePropType<Color>(Object),
+    required: true,
+  },
+  vertical: {
+    type: Boolean,
+    default: false,
+  },
+} as const)
+
+export type AlphaSliderProps = ExtractPropTypes<typeof alphaSliderProps>
